+++ src/main/java/org/example/Main.java
@@ -1,3 +1,5 @@
+package org.example;
+
 public class Main {
     public static void main(String[] args) {
