+++ src/main/java/org/example/entities/Cliente.java
@@ -1,3 +1,5 @@
+package org.example.entities;
+
 import java.time.LocalDate;
 import java.util.HashSet;
 import java.util.Set;
