+++ src/main/java/org/example/entities/Domicilio.java
@@ -1,3 +1,5 @@
+package org.example.entities;
+
 public class Domicilio {
     private String calle;
     private Integer numero;
