+++ src/main/java/org/example/entities/Localidad.java
@@ -1,3 +1,5 @@
+package org.example.entities;
+
 public class Localidad {
     private String nombre;
     private Provincia provincia;
