+++ src/main/java/org/example/entities/Usuario.java
@@ -1,3 +1,5 @@
+package org.example.entities;
+
 public class Usuario {
     private String auth0Id;
     private String username;
