+++ src/main/java/org/example/entities/enums/Rol.java
@@ -1,3 +1,5 @@
+package org.example.entities.enums;
+
 public enum Rol {
     ADMIN,
     EMPLEADO,
