+++ src/main/java/org/example/entities/enums/TipoEnvio.java
@@ -1,3 +1,5 @@
+package org.example.entities.enums;
+
 public enum TipoEnvio {
     DELIVERY,
     TAKEAWAY
