+++ src/main/java/org/example/entities/enums/TipoPromocion.java
@@ -1,3 +1,5 @@
+package org.example.entities.enums;
+
 public enum TipoPromocion {
     HAPPYHOUR,PROMOCION1;
 }
